refactor&feat: add sealed decorator and uniformize filename casing
