+++ frontend/src/components/Buttons/Playback/PlayPauseButton.vue
@@ -15,7 +15,7 @@ import IMdiExclamation from 'virtual:icons/mdi/exclamation';
 import IMdiPauseCircleOutline from 'virtual:icons/mdi/pause-circle-outline';
 import IMdiPlayCircleOutline from 'virtual:icons/mdi/play-circle-outline';
 import { computed } from 'vue';
-import { PlaybackStatus, playbackManager } from '@/store/playbackManager';
+import { PlaybackStatus, playbackManager } from '@/store/playback-manager';
 
 defineProps<{ size?: string }>();
