+++ frontend/src/components/Buttons/Playback/RepeatButton.vue
@@ -14,7 +14,7 @@
 import IMdiRepeat from 'virtual:icons/mdi/repeat';
 import IMdiRepeatOnce from 'virtual:icons/mdi/repeat-once';
 import { computed } from 'vue';
-import { RepeatMode, playbackManager } from '@/store/playbackManager';
+import { RepeatMode, playbackManager } from '@/store/playback-manager';
 
 defineProps<{ size?: string }>();
