+++ frontend/src/components/Layout/AppBar/Buttons/TaskManagerButton.vue
@@ -49,7 +49,7 @@
 <script setup lang="ts">
 import { computed, ref, watch } from 'vue';
 import { useI18n } from 'vue-i18n';
-import { taskManager, TaskType } from '@/store/taskManager';
+import { taskManager, TaskType } from '@/store/task-manager';
 
 interface TaskInfo {
   progress: undefined | number;
