+++ frontend/src/components/Layout/TimeSlider.vue
@@ -21,7 +21,7 @@
 
 <script setup lang="ts">
 import { computed, ref } from 'vue';
-import { playbackManager } from '@/store/playbackManager';
+import { playbackManager } from '@/store/playback-manager';
 import { formatTime } from '@/utils/time';
 
 const currentInput = ref(0);
