+++ frontend/src/components/Playback/DraggableQueue.vue
@@ -45,7 +45,7 @@ import type { BaseItemDto } from '@jellyfin/sdk/lib/generated-client';
 import Sortable from 'sortablejs';
 import { onBeforeUnmount, shallowRef, watch } from 'vue';
 import { isNumber } from '@/utils/validation';
-import { playbackManager } from '@/store/playbackManager';
+import { playbackManager } from '@/store/playback-manager';
 
 let sortable: Sortable | undefined;
 const container = shallowRef<HTMLSpanElement>();
