+++ frontend/src/components/Playback/PiPVideoPlayer.vue
@@ -74,8 +74,8 @@
 <script setup lang="ts">
 import { useMagicKeys, whenever } from '@vueuse/core';
 import { onBeforeUnmount, onMounted } from 'vue';
-import { playbackManager } from '@/store/playbackManager';
-import { playerElement } from '@/store/playerElement';
+import { playbackManager } from '@/store/playback-manager';
+import { playerElement } from '@/store/player-element';
 
 const keys = useMagicKeys();
