+++ frontend/src/components/Playback/PlayerElement.vue
@@ -36,8 +36,8 @@ import {
   mediaElementRef,
   mediaWebAudio
 } from '@/store';
-import { playbackManager } from '@/store/playbackManager';
-import { playerElement } from '@/store/playerElement';
+import { playbackManager } from '@/store/playback-manager';
+import { playerElement } from '@/store/player-element';
 import { getImageInfo } from '@/utils/images';
 import { isNil } from '@/utils/validation';
