+++ frontend/src/components/Playback/UpNext.vue
@@ -69,7 +69,7 @@
 
 <script setup lang="ts">
 import { computed, ref, watch } from 'vue';
-import { playbackManager } from '@/store/playbackManager';
+import { playbackManager } from '@/store/playback-manager';
 import { getEndsAtTime, getRuntimeTime } from '@/utils/time';
 
 const emit = defineEmits<{
