+++ frontend/src/components/System/LocaleSwitcher.vue
@@ -29,7 +29,7 @@
 
 <script setup lang="ts">
 import { useI18n } from 'vue-i18n';
-import { clientSettings } from '@/store/clientSettings';
+import { clientSettings } from '@/store/client-settings';
 import { getLocaleNativeName } from '@/utils/i18n';
 
 defineProps<{
