+++ frontend/src/layouts/default.vue
@@ -21,8 +21,8 @@ import type { BaseItemDto } from '@jellyfin/sdk/lib/generated-client';
 import { computed, onBeforeMount, onUnmounted, provide, ref, watch } from 'vue';
 import { useDisplay } from 'vuetify';
 import type { DrawerItem } from '@/components/Layout/Navigation/NavigationDrawer.vue';
-import { playbackManager } from '@/store/playbackManager';
-import { playerElement } from '@/store/playerElement';
+import { playbackManager } from '@/store/playback-manager';
+import { playerElement } from '@/store/player-element';
 import { fetchIndexPage, getLibraryIcon } from '@/utils/items';
 
 const display = useDisplay();
