+++ frontend/src/plugins/remote/index.ts
@@ -10,9 +10,10 @@ import type { App } from 'vue';
 import RemotePluginAuthInstance from './auth';
 import RemotePluginSDKInstance from './sdk';
 import RemotePluginSocketInstance from './socket';
-import { isNil } from '@/utils/validation';
+import { isNil, sealed } from '@/utils/validation';
 import { getJSONConfig } from '@/utils/external-config';
 
+@sealed
 class RemotePlugin {
   public readonly auth = RemotePluginAuthInstance;
   public readonly sdk = RemotePluginSDKInstance;
